+++ apps/avifdec.c
@@ -347,6 +347,18 @@ int main(int argc, char * argv[])
     printf("Image details:\n");
     avifImageDump(decoder->image, 0, 0, decoder->progressiveState);
 
+    if (decoder->image->transformFlags & AVIF_TRANSFORM_CLAP) {
+        avifCropRect cropRect;
+        if (!avifCropRectFromCleanApertureBox(&cropRect,
+                                              &decoder->image->clap,
+                                              decoder->image->width,
+                                              decoder->image->height,
+                                              &decoder->diag)) {
+            // Should happen only if AVIF_STRICT_CLAP_VALID is disabled.
+            fprintf(stderr, "Warning: Invalid Clean Aperture values\n");
+        }
+    }
+
     if (ignoreICC && (decoder->image->icc.size > 0)) {
         printf("[--ignore-icc] Discarding ICC profile.\n");
         // This cannot fail.