+++ apps/shared/avifjpeg.c
@@ -1300,7 +1300,7 @@ avifBool avifJPEGWrite(const char * outputFilename, const avifImage * avif, int 
     if (avif->transformFlags & AVIF_TRANSFORM_CLAP) {
         avifCropRect cropRect;
         avifDiagnostics diag;
-        if (avifCropRectConvertCleanApertureBox(&cropRect, &avif->clap, avif->width, avif->height, avif->yuvFormat, &diag) &&
+        if (avifCropRectFromCleanApertureBox(&cropRect, &avif->clap, avif->width, avif->height, &diag) &&
             (cropRect.x != 0 || cropRect.y != 0 || cropRect.width != avif->width || cropRect.height != avif->height)) {
             // TODO: https://github.com/AOMediaCodec/libavif/issues/2427 - Implement.
             fprintf(stderr,