+++ src/avif.c
@@ -749,19 +749,8 @@ static avifBool overflowsInt32(int64_t x)
     return (x < INT32_MIN) || (x > INT32_MAX);
 }
 
-static avifBool avifCropRectIsValid(const avifCropRect * cropRect, uint32_t imageW, uint32_t imageH, avifPixelFormat yuvFormat, avifDiagnostics * diag)
-
+static avifBool avifCropRectIsValid(const avifCropRect * cropRect, uint32_t imageW, uint32_t imageH, avifDiagnostics * diag)
 {
-    // ISO/IEC 23000-22:2019/Amd. 2:2021, Section 7.3.6.7:
-    //   The clean aperture property is restricted according to the chroma
-    //   sampling format of the input image (4:4:4, 4:2:2:, 4:2:0, or 4:0:0) as
-    //   follows:
-    //   ...
-    //   - If chroma is subsampled horizontally (i.e., 4:2:2 and 4:2:0), the
-    //     leftmost pixel of the clean aperture shall be even numbers;
-    //   - If chroma is subsampled vertically (i.e., 4:2:0), the topmost line
-    //     of the clean aperture shall be even numbers.
-
     if ((cropRect->width == 0) || (cropRect->height == 0)) {
         avifDiagnosticsPrintf(diag, "[Strict] crop rect width and height must be nonzero");
         return AVIF_FALSE;
@@ -771,28 +760,14 @@ static avifBool avifCropRectIsValid(const avifCropRect * cropRect, uint32_t imag
         avifDiagnosticsPrintf(diag, "[Strict] crop rect is out of the image's bounds");
         return AVIF_FALSE;
     }
-
-    if ((yuvFormat == AVIF_PIXEL_FORMAT_YUV420) || (yuvFormat == AVIF_PIXEL_FORMAT_YUV422)) {
-        if ((cropRect->x % 2) != 0) {
-            avifDiagnosticsPrintf(diag, "[Strict] crop rect X offset must be even due to this image's YUV subsampling");
-            return AVIF_FALSE;
-        }
-    }
-    if (yuvFormat == AVIF_PIXEL_FORMAT_YUV420) {
-        if ((cropRect->y % 2) != 0) {
-            avifDiagnosticsPrintf(diag, "[Strict] crop rect Y offset must be even due to this image's YUV subsampling");
-            return AVIF_FALSE;
-        }
-    }
     return AVIF_TRUE;
 }
 
-avifBool avifCropRectConvertCleanApertureBox(avifCropRect * cropRect,
-                                             const avifCleanApertureBox * clap,
-                                             uint32_t imageW,
-                                             uint32_t imageH,
-                                             avifPixelFormat yuvFormat,
-                                             avifDiagnostics * diag)
+avifBool avifCropRectFromCleanApertureBox(avifCropRect * cropRect,
+                                          const avifCleanApertureBox * clap,
+                                          uint32_t imageW,
+                                          uint32_t imageH,
+                                          avifDiagnostics * diag)
 {
     avifDiagnosticsClearError(diag);
 
@@ -819,9 +794,6 @@ avifBool avifCropRectConvertCleanApertureBox(avifCropRect * cropRect,
     }
 
     // ISO/IEC 23000-22:2019/Amd. 2:2021, Section 7.3.6.7:
-    //   The clean aperture property is restricted according to the chroma
-    //   sampling format of the input image (4:4:4, 4:2:2:, 4:2:0, or 4:0:0) as
-    //   follows:
     //   - cleanApertureWidth and cleanApertureHeight shall be integers;
     //   - The leftmost pixel and the topmost line of the clean aperture as
     //     defined in ISO/IEC 14496-12:2020, Section 12.1.4.1 shall be integers;
@@ -898,19 +870,18 @@ avifBool avifCropRectConvertCleanApertureBox(avifCropRect * cropRect,
     cropRect->y = (uint32_t)(cropY.n / cropY.d);
     cropRect->width = (uint32_t)clapW;
     cropRect->height = (uint32_t)clapH;
-    return avifCropRectIsValid(cropRect, imageW, imageH, yuvFormat, diag);
+    return avifCropRectIsValid(cropRect, imageW, imageH, diag);
 }
 
-avifBool avifCleanApertureBoxConvertCropRect(avifCleanApertureBox * clap,
-                                             const avifCropRect * cropRect,
-                                             uint32_t imageW,
-                                             uint32_t imageH,
-                                             avifPixelFormat yuvFormat,
-                                             avifDiagnostics * diag)
+avifBool avifCleanApertureBoxFromCropRect(avifCleanApertureBox * clap,
+                                          const avifCropRect * cropRect,
+                                          uint32_t imageW,
+                                          uint32_t imageH,
+                                          avifDiagnostics * diag)
 {
     avifDiagnosticsClearError(diag);
 
-    if (!avifCropRectIsValid(cropRect, imageW, imageH, yuvFormat, diag)) {
+    if (!avifCropRectIsValid(cropRect, imageW, imageH, diag)) {
         return AVIF_FALSE;
     }
 
@@ -963,6 +934,68 @@ avifBool avifCleanApertureBoxConvertCropRect(avifCleanApertureBox * clap,
     clap->vertOffN = vertOff.n;
     clap->vertOffD = vertOff.d;
     return AVIF_TRUE;
+}
+
+avifBool avifCropRectRequiresUpsampling(const avifCropRect * cropRect, avifPixelFormat yuvFormat)
+{
+    // ISO/IEC 23000-22:2024 FDIS, Section 7.3.6.7:
+    // - If any of the following conditions hold true, the image is first implicitly upsampled to 4:4:4:
+    //   - chroma is subsampled horizontally (i.e., 4:2:2 and 4:2:0) and cleanApertureWidth is odd
+    //   - chroma is subsampled horizontally (i.e., 4:2:2 and 4:2:0) and left-most pixel is on an odd position
+    //   - chroma is subsampled vertically (i.e., 4:2:0) and cleanApertureHeight is odd
+    //   - chroma is subsampled vertically (i.e., 4:2:0) and topmost line is on an odd position
+
+    // AV1 supports odd dimensions with chroma subsampling in those directions, so only look for x and y.
+    return ((yuvFormat == AVIF_PIXEL_FORMAT_YUV420 || yuvFormat == AVIF_PIXEL_FORMAT_YUV422) && (cropRect->x % 2)) ||
+           (yuvFormat == AVIF_PIXEL_FORMAT_YUV420 && (cropRect->y % 2));
+}
+
+avifBool avifCropRectConvertCleanApertureBox(avifCropRect * cropRect,
+                                             const avifCleanApertureBox * clap,
+                                             uint32_t imageW,
+                                             uint32_t imageH,
+                                             avifPixelFormat yuvFormat,
+                                             avifDiagnostics * diag)
+{
+    if (!avifCropRectFromCleanApertureBox(cropRect, clap, imageW, imageH, diag)) {
+        return AVIF_FALSE;
+    }
+    // Keep the same pre-deprecation behavior.
+
+    // ISO/IEC 23000-22:2019/Amd. 2:2021, Section 7.3.6.7:
+    //   - If chroma is subsampled horizontally (i.e., 4:2:2 and 4:2:0),
+    //     the leftmost pixel of the clean aperture shall be even numbers;
+    //   - If chroma is subsampled vertically (i.e., 4:2:0),
+    //     the topmost line of the clean aperture shall be even numbers.
+
+    if (avifCropRectRequiresUpsampling(cropRect, yuvFormat)) {
+        avifDiagnosticsPrintf(diag, "[Strict] crop rect X and Y offsets must be even due to this image's YUV subsampling");
+        return AVIF_FALSE;
+    }
+    return AVIF_TRUE;
+}
+
+avifBool avifCleanApertureBoxConvertCropRect(avifCleanApertureBox * clap,
+                                             const avifCropRect * cropRect,
+                                             uint32_t imageW,
+                                             uint32_t imageH,
+                                             avifPixelFormat yuvFormat,
+                                             avifDiagnostics * diag)
+{
+    // Keep the same pre-deprecation behavior.
+
+    // ISO/IEC 23000-22:2019/Amd. 2:2021, Section 7.3.6.7:
+    //   - If chroma is subsampled horizontally (i.e., 4:2:2 and 4:2:0),
+    //     the leftmost pixel of the clean aperture shall be even numbers;
+    //   - If chroma is subsampled vertically (i.e., 4:2:0),
+    //     the topmost line of the clean aperture shall be even numbers.
+
+    if (avifCropRectRequiresUpsampling(cropRect, yuvFormat)) {
+        avifDiagnosticsPrintf(diag, "[Strict] crop rect X and Y offsets must be even due to this image's YUV subsampling");
+        return AVIF_FALSE;
+    }
+
+    return avifCleanApertureBoxFromCropRect(clap, cropRect, imageW, imageH, diag);
 }
 
 // ---------------------------------------------------------------------------