+++ tests/gtest/avifclaptest.cc
@@ -82,14 +82,13 @@ using InvalidClapPropertyTest =
 INSTANTIATE_TEST_SUITE_P(Parameterized, InvalidClapPropertyTest,
                          ::testing::ValuesIn(kInvalidClapPropertyTestParams));
 
-// Negative tests for the avifCropRectConvertCleanApertureBox() function.
+// Negative tests for the avifCropRectFromCleanApertureBox() function.
 TEST_P(InvalidClapPropertyTest, ValidateClapProperty) {
   const InvalidClapPropertyParam& param = GetParam();
   avifCropRect crop_rect;
   avifDiagnostics diag;
-  EXPECT_FALSE(avifCropRectConvertCleanApertureBox(&crop_rect, &param.clap,
-                                                   param.width, param.height,
-                                                   param.yuv_format, &diag));
+  EXPECT_FALSE(avifCropRectFromCleanApertureBox(
+      &crop_rect, &param.clap, param.width, param.height, &diag));
 }
 
 struct ValidClapPropertyParam {
@@ -99,6 +98,7 @@ struct ValidClapPropertyParam {
   avifCleanApertureBox clap;
 
   avifCropRect expected_crop_rect;
+  bool expected_upsample_before_cropping;
 };
 
 constexpr ValidClapPropertyParam kValidClapPropertyTestParams[] = {
@@ -110,7 +110,8 @@ constexpr ValidClapPropertyParam kValidClapPropertyTestParams[] = {
      160,
      AVIF_PIXEL_FORMAT_YUV420,
      {96, 1, 132, 1, 0, 1, 0, 1},
-     {12, 14, 96, 132}},
+     {12, 14, 96, 132},
+     false},
     // pcX = -30 + (120 - 1)/2 = 29.5
     // pcY = -40 + (160 - 1)/2 = 39.5
     // leftmost = 29.5 - (60 - 1)/2 = 0
@@ -120,7 +121,8 @@ constexpr ValidClapPropertyParam kValidClapPropertyTestParams[] = {
      AVIF_PIXEL_FORMAT_YUV420,
      {60, 1, 80, 1, static_cast<uint32_t>(-30), 1, static_cast<uint32_t>(-40),
       1},
-     {0, 0, 60, 80}},
+     {0, 0, 60, 80},
+     false},
     // pcX = -1/2 + (100 - 1)/2 = 49
     // pcY = -1/2 + (100 - 1)/2 = 49
     // leftmost = 49 - (99 - 1)/2 = 0
@@ -129,7 +131,18 @@ constexpr ValidClapPropertyParam kValidClapPropertyTestParams[] = {
      100,
      AVIF_PIXEL_FORMAT_YUV420,
      {99, 1, 99, 1, static_cast<uint32_t>(-1), 2, static_cast<uint32_t>(-1), 2},
-     {0, 0, 99, 99}},
+     {0, 0, 99, 99},
+     false},
+    // pcX = 1/2 + (100 - 1)/2 = 50
+    // pcY = 1/2 + (100 - 1)/2 = 50
+    // leftmost = 50 - (99 - 1)/2 = 1
+    // topmost = 50 - (99 - 1)/2 = 1
+    {100,
+     100,
+     AVIF_PIXEL_FORMAT_YUV420,
+     {99, 1, 99, 1, 1, 2, 1, 2},
+     {1, 1, 99, 99},
+     true},
 };
 
 using ValidClapPropertyTest = ::testing::TestWithParam<ValidClapPropertyParam>;
@@ -137,19 +150,63 @@ using ValidClapPropertyTest = ::testing::TestWithParam<ValidClapPropertyParam>;
 INSTANTIATE_TEST_SUITE_P(Parameterized, ValidClapPropertyTest,
                          ::testing::ValuesIn(kValidClapPropertyTestParams));
 
-// Positive tests for the avifCropRectConvertCleanApertureBox() function.
+// Positive tests for the avifCropRectFromCleanApertureBox() and
+// avifCleanApertureBoxFromCropRect() functions.
 TEST_P(ValidClapPropertyTest, ValidateClapProperty) {
   const ValidClapPropertyParam& param = GetParam();
   avifCropRect crop_rect;
   avifDiagnostics diag;
-  EXPECT_TRUE(avifCropRectConvertCleanApertureBox(&crop_rect, &param.clap,
-                                                  param.width, param.height,
-                                                  param.yuv_format, &diag))
+  ASSERT_TRUE(avifCropRectFromCleanApertureBox(
+      &crop_rect, &param.clap, param.width, param.height, &diag))
       << diag.error;
+  const avifBool upsample_before_cropping =
+      avifCropRectRequiresUpsampling(&crop_rect, param.yuv_format);
   EXPECT_EQ(crop_rect.x, param.expected_crop_rect.x);
   EXPECT_EQ(crop_rect.y, param.expected_crop_rect.y);
   EXPECT_EQ(crop_rect.width, param.expected_crop_rect.width);
   EXPECT_EQ(crop_rect.height, param.expected_crop_rect.height);
+  EXPECT_EQ(upsample_before_cropping, param.expected_upsample_before_cropping);
+
+  // Deprecated function coverage.
+  avifBool success = avifCropRectConvertCleanApertureBox(
+      &crop_rect, &param.clap, param.width, param.height, param.yuv_format,
+      &diag);
+  EXPECT_EQ(success, !upsample_before_cropping);
+  if (success) {
+    EXPECT_EQ(crop_rect.x, param.expected_crop_rect.x);
+    EXPECT_EQ(crop_rect.y, param.expected_crop_rect.y);
+    EXPECT_EQ(crop_rect.width, param.expected_crop_rect.width);
+    EXPECT_EQ(crop_rect.height, param.expected_crop_rect.height);
+  }
+
+  avifCleanApertureBox clap;
+  ASSERT_TRUE(avifCleanApertureBoxFromCropRect(
+      &clap, &param.expected_crop_rect, param.width, param.height, &diag))
+      << diag.error;
+  EXPECT_EQ(clap.widthN, param.clap.widthN);
+  EXPECT_EQ(clap.widthD, param.clap.widthD);
+  EXPECT_EQ(clap.heightN, param.clap.heightN);
+  EXPECT_EQ(clap.heightD, param.clap.heightD);
+  EXPECT_EQ(clap.horizOffN, param.clap.horizOffN);
+  EXPECT_EQ(clap.horizOffD, param.clap.horizOffD);
+  EXPECT_EQ(clap.vertOffN, param.clap.vertOffN);
+  EXPECT_EQ(clap.vertOffD, param.clap.vertOffD);
+
+  // Deprecated function coverage.
+  success = avifCleanApertureBoxConvertCropRect(
+      &clap, &param.expected_crop_rect, param.width, param.height,
+      param.yuv_format, &diag);
+  EXPECT_EQ(success, !upsample_before_cropping);
+  if (success) {
+    EXPECT_EQ(clap.widthN, param.clap.widthN);
+    EXPECT_EQ(clap.widthD, param.clap.widthD);
+    EXPECT_EQ(clap.heightN, param.clap.heightN);
+    EXPECT_EQ(clap.heightD, param.clap.heightD);
+    EXPECT_EQ(clap.horizOffN, param.clap.horizOffN);
+    EXPECT_EQ(clap.horizOffD, param.clap.horizOffD);
+    EXPECT_EQ(clap.vertOffN, param.clap.vertOffN);
+    EXPECT_EQ(clap.vertOffD, param.clap.vertOffD);
+  }
 }
 
 }  // namespace